ERESI 0.83a has the following additions:
- Fixed eresi-config scripts so that more testsuite build with them

- Fixed e2dbg to aoid using s*printf functions at initialization time as these could use malloc on newer glibc whose symbol is not resolved yet.

- Fixed revm_expr_extend() to add fields to expressions on the fly during transmation (to be cont)

- Added revm_expr_contex_t so that many revm expr engine API is more modular and API take less parameters

- Corrected a one-byte heap overflow in the revm expr engine (fieldname dup should have allocated +1 for NUL byte)

FOUND NEW BUGS: 

- There is a new section type .gnu.hash that is now replacing legacy .hash - need libelfsh support
- e2dbg fails to handle SIGTRAP in-process as before even on linux/x86, maybe a mprotect-related problem, reenabled sigsegv handling

More to come.

+++ librevm/lang/lookup.c
@@ -450,7 +450,7 @@ revmexpr_t		*revm_lookup_param(char *param, u_char existing)
   char			**keys;
   int			keynbr;
   char			*parm;
-  unsigned int		sepnbr;
+  unsigned short	sepnbr;
   int			index;
   revmobj_t		*res;
   revmexpr_t		*expr;
@@ -471,8 +471,7 @@ revmexpr_t		*revm_lookup_param(char *param, u_char existing)
   for (sepnbr = 0, parm = param; *parm; parm++)
     if (*parm == REVM_SEP[0] || *parm == ':' || *parm == '%')
       {
-	if (parm > param && parm[-1] == '[')
-	  continue;
+	if (parm > param && parm[-1] == '[') continue;
 	sepnbr++;
       }
